requirements.txt
openai
tiktoken
llama-cpp-python